+++ tools/onnx/onnx2ncnn.cpp
@@ -12,45 +12,16 @@
 // CONDITIONS OF ANY KIND, either express or implied. See the License for the
 // specific language governing permissions and limitations under the License.
 
-#include "onnx.pb.h"
+#include "onnx2ncnn.h"
 
 #include <algorithm>
 #include <float.h>
-#include <fstream>
-#include <google/protobuf/io/coded_stream.h>
-#include <google/protobuf/io/zero_copy_stream_impl.h>
 #include <google/protobuf/message.h>
 #include <google/protobuf/text_format.h>
 #include <iostream>
 #include <limits.h>
 #include <limits>
 #include <set>
-#include <stdio.h>
-
-static bool read_proto_from_binary(const char* filepath, onnx::ModelProto* message)
-{
-    std::ifstream fs(filepath, std::ifstream::in | std::ifstream::binary);
-    if (!fs.is_open())
-    {
-        fprintf(stderr, "open failed %s\n", filepath);
-        return false;
-    }
-
-    google::protobuf::io::IstreamInputStream input(&fs);
-    google::protobuf::io::CodedInputStream codedstr(&input);
-
-#if GOOGLE_PROTOBUF_VERSION >= 3011000
-    codedstr.SetTotalBytesLimit(INT_MAX);
-#else
-    codedstr.SetTotalBytesLimit(INT_MAX, INT_MAX / 2);
-#endif
-
-    bool success = message->ParseFromCodedStream(&codedstr);
-
-    fs.close();
-
-    return success;
-}
 
 static std::vector<int> get_node_attr_ai(const onnx::NodeProto& node, const char* key)
 {
@@ -2083,25 +2054,7 @@ static void fuse_lstm_gru_rnn(onnx::GraphProto* mutable_graph, std::map<std::str
     }
 }
 
-int main(int argc, char** argv)
-{
-    const char* onnxpb = argv[1];
-    const char* ncnn_prototxt = argc >= 4 ? argv[2] : "ncnn.param";
-    const char* ncnn_modelbin = argc >= 4 ? argv[3] : "ncnn.bin";
-
-    onnx::ModelProto model;
-
-    // load
-    bool s1 = read_proto_from_binary(onnxpb, &model);
-    if (!s1)
-    {
-        fprintf(stderr, "read_proto_from_binary failed\n");
-        return -1;
-    }
-
-    FILE* pp = fopen(ncnn_prototxt, "wb");
-    FILE* bp = fopen(ncnn_modelbin, "wb");
-
+void convert(onnx::ModelProto model, FILE* pp, FILE* bp) {
     // magic
     fprintf(pp, "7767517\n");
 
@@ -4670,9 +4623,4 @@ int main(int argc, char** argv)
             }
         }
     }
-
-    fclose(pp);
-    fclose(bp);
-
-    return 0;
 }
